+++ test/tests/date.js
@@ -9,7 +9,7 @@ package('Date', function () {
     // Imaginary locale to test locale switching
     testAddLocale('fo', {
       units: 'do,re,mi,fa,so,la,ti,do',
-      months: 'do,re,mi,fa,so,la,ti,do',
+      months: 'Do,Re,Mi,Fa,So,La,Ti,Do',
       parse: [
         '{year}kupo',
         '{month}mofo'
@@ -765,10 +765,11 @@ package('Date', function () {
 
     var d = run(testCreateUTCDate('2000-02-14'), 'setUTC', [true]);
 
-    equal(run(d, 'format'), 'February 14, 2000 12:00am', 'fomratting monday');
-    equal(run(d, 'full'), 'Monday February 14, 2000 12:00:00am', 'full format');
-    equal(run(d, 'long'), 'February 14, 2000 12:00am', 'long format');
-    equal(run(d, 'short'), 'February 14, 2000', 'short format');
+    equal(run(d, 'format'), 'February 14, 2000 12:00 AM', 'formatting monday');
+    equal(run(d, 'short'), '02/14/2000', 'short format');
+    equal(run(d, 'medium'), 'February 14, 2000', 'medium format');
+    equal(run(d, 'long'), 'February 14, 2000 12:00 AM', 'long format');
+    equal(run(d, 'full'), 'Monday, February 14, 2000 12:00 AM', 'full format');
 
     // UTC flag is now deprecated in comparison methods so instead we
     // need to run through Date#create.
@@ -1472,13 +1473,10 @@ package('Date', function () {
   method('format', function() {
     var d = new Date('August 5, 2010 13:45:02');
 
-    test(d, 'August 5, 2010 1:45pm', 'no arguments is standard format with no time');
+    test(d, 'August 5, 2010 1:45 PM', 'no arguments is standard format with no time');
     test(d, ['{ms}'], '0', 'ms');
-    test(d, ['{milliseconds}'], '0', 'milliseconds');
-    test(d, ['{f}'], '0', 'f');
-    test(d, ['{ff}'], '00', 'ff');
-    test(d, ['{fff}'], '000', 'fff');
-    test(d, ['{ffff}'], '0000', 'ffff');
+    test(d, ['{S}'], '0', 'S');
+    test(d, ['{SSS}'], '000', 'SSS');
     test(d, ['{s}'], '2', 's');
     test(d, ['{ss}'], '02', 'ss');
     test(d, ['{seconds}'], '2', 'seconds');
@@ -1496,27 +1494,106 @@ package('Date', function () {
     test(d, ['{dd}'], '05', 'dd');
     test(d, ['{date}'], '5', 'date');
     test(d, ['{day}'], '5', 'days');
+    test(d, ['{do}'], '5th', 'Day ordinal');
     test(d, ['{dow}'], 'thu', 'dow');
     test(d, ['{Dow}'], 'Thu', 'Dow');
-    test(d, ['{do}'], 'th', 'do');
-    test(d, ['{Do}'], 'Th', 'Do');
+    test(d, ['{e}'], '4', 'Day of week as number');
+    test(d, ['{eo}'], '4th', 'Day of week as ordinal number');
     test(d, ['{weekday}'], 'thursday', 'weekday');
     test(d, ['{Weekday}'], 'Thursday', 'Weekday');
+    test(d, ['{D}'], '217', 'Day of the year');
+    test(d, ['{DDD}'], '217', 'Day of the year padded');
+    test(d, ['{gg}'], '10', '2 digit week year');
+    test(d, ['{gggg}'], '2010', '4 digit week year');
+    test(d, ['{GG}'], '10', '2 digit ISO week year');
+    test(d, ['{GGGG}'], '2010', '4 digit ISO week year');
     test(d, ['{M}'], '8', 'M');
     test(d, ['{MM}'], '08', 'MM');
     test(d, ['{month}'], 'august', 'month');
     test(d, ['{Mon}'], 'Aug', 'Mon');
     test(d, ['{Month}'], 'August', 'Month');
+    test(d, ['{Mo}'], '8th', 'Month ordinal');
     test(d, ['{yy}'], '10', 'yy');
     test(d, ['{yyyy}'], '2010', 'yyyy');
     test(d, ['{year}'], '2010', 'year');
+    test(d, ['{Q}'], '3', 'Quarter');
     test(d, ['{t}'], 'p', 't');
     test(d, ['{T}'], 'P', 'T');
     test(d, ['{tt}'], 'pm', 'tt');
     test(d, ['{TT}'], 'PM', 'TT');
-    test(d, ['{ord}'], '5th', 'ord');
-    equal(!!run(d, 'format', ['{Z}']).match(testGetTimezoneHours(d)), true, 'Z');
-    equal(!!run(d, 'format', ['{ZZ}']).match(testGetTimezoneHours(d)), true, 'Z');
+    test(d, ['{w}'], '32', 'locale week number (English)');
+    test(d, ['{ww}'], '32', 'locale week number padded (English)');
+    test(d, ['{wo}'], '32nd', 'locale week number ordinal (English)');
+    test(d, ['{W}'], '31', 'ISO week number');
+    test(d, ['{WW}'], '31', 'ISO week number padded');
+    test(d, ['{Wo}'], '31st', 'ISO week number ordinal');
+    test(d, ['{X}'], '1280983502', 'Unix timestamp');
+    test(d, ['{x}'], '1280983502000', 'Unix millisecond timestamp');
+
+    test(d, ['{Z}'],  getExpectedTimezoneOffset(d, true), 'Z');
+    test(d, ['{ZZ}'], getExpectedTimezoneOffset(d), 'ZZ');
+
+    // Not all environments provide that so just make sure it returns the abbreviation or nothing.
+    equal(/\w{3}|^$/.test(run(d, 'format', ['{z}'])), true, 'Timezone abbreviation');
+
+    test(new Date('January 4, 2010'), ['{D}'], '4', 'Day of the year');
+    test(new Date('January 4, 2010'), ['{DDD}'], '004', 'Day of the year padded');
+
+    test(new Date('January 3, 2010'), ['{W}'], '53', 'ISO week number | Jan 3 2010');
+    test(new Date('January 3, 2010'), ['{WW}'], '53', 'ISO week number padded | Jan 3 2010');
+    test(new Date('January 3, 2010'), ['{Wo}'], '53rd', 'ISO week number ordinal | Jan 3 2010');
+    test(new Date('January 4, 2010'), ['{W}'], '1', 'ISO week number | Jan 4 2010');
+    test(new Date('January 4, 2010'), ['{WW}'], '01', 'ISO week number padded | Jan 4 2010');
+    test(new Date('January 4, 2010'), ['{Wo}'], '1st', 'ISO week number ordinal | Jan 4 2010');
+
+    test(new Date('December 26, 2009'), ['{w}'], '52', 'locale week number | Dec 26 2009');
+    test(new Date('December 26, 2009'), ['{ww}'], '52', 'locale week number padded | Dec 26 2009');
+    test(new Date('December 26, 2009'), ['{wo}'], '52nd', 'locale week number ordinal | Dec 26 2009');
+    test(new Date('December 27, 2009'), ['{w}'], '1', 'locale week number | Dec 27 2009');
+    test(new Date('December 27, 2009'), ['{ww}'], '01', 'locale week number padded | Dec 27 2009');
+    test(new Date('December 27, 2009'), ['{wo}'], '1st', 'locale week number ordinal | Dec 27 2009');
+
+    test(new Date('January 3, 2010'), ['{w}', 'en-GB'], '53', 'locale week number | Jan 3 2010 | UK');
+    test(new Date('January 3, 2010'), ['{ww}', 'en-GB'], '53', 'locale week number padded | Jan 3 2010 | UK');
+    test(new Date('January 3, 2010'), ['{wo}', 'en-GB'], '53rd', 'locale week number ordinal | Jan 3 2010 | UK');
+    test(new Date('January 4, 2010'), ['{w}', 'en-GB'], '1', 'locale week number | Jan 4 2010 | UK');
+    test(new Date('January 4, 2010'), ['{ww}', 'en-GB'], '01', 'locale week number padded | Jan 4 2010 | UK');
+    test(new Date('January 4, 2010'), ['{wo}', 'en-GB'], '1st', 'locale week number ordinal | Jan 4 2010 | UK');
+
+    test(new Date('December 26, 2009'), ['{gg}'], '09', '2 digit week year');
+    test(new Date('December 26, 2009'), ['{gggg}'], '2009', '4 digit week year');
+    test(new Date('December 27, 2009'), ['{gg}'], '10', '2 digit week year | next');
+    test(new Date('December 27, 2009'), ['{gggg}'], '2010', '4 digit week year | next');
+
+    test(new Date('December 28, 2008'), ['{gg}', 'en-GB'], '08', '2 digit week year | UK');
+    test(new Date('December 28, 2008'), ['{gggg}', 'en-GB'], '2008', '4 digit week year | UK');
+    test(new Date('December 29, 2008'), ['{gg}', 'en-GB'], '09', '2 digit week year | next | UK');
+    test(new Date('December 29, 2008'), ['{gggg}', 'en-GB'], '2009', '4 digit week year | next | UK');
+    test(new Date('January 3, 2010'),   ['{gg}', 'en-GB'], '09', '2 digit week year | prev | UK');
+    test(new Date('January 3, 2010'),   ['{gggg}', 'en-GB'], '2009', '4 digit week year | prev | UK');
+
+    test(new Date('December 28, 2008'), ['{GG}'], '08', '2 digit week year | ISO8601');
+    test(new Date('December 28, 2008'), ['{GGGG}'], '2008', '4 digit week year | ISO8601');
+    test(new Date('December 29, 2008'), ['{GG}'], '09', '2 digit week year | next | ISO8601');
+    test(new Date('December 29, 2008'), ['{GGGG}'], '2009', '4 digit week year | next | ISO8601');
+    test(new Date('January 3, 2010'),   ['{GG}'], '09', '2 digit week year | prev | ISO8601');
+    test(new Date('January 3, 2010'),   ['{GGGG}'], '2009', '4 digit week year | prev | ISO8601');
+
+    test(new Date(-60000000000000), ['{yy}'], '68', 'yy in year 0068 should be 2 digits');
+    test(new Date(-60000000000000), ['{yyyy}'], '0068', 'yyyy in year 0068 should be 4 digits');
+
+    test(new Date(2010,  0), ['{Q}'], '1', 'Jan is Q1');
+    test(new Date(2010,  1), ['{Q}'], '1', 'Feb is Q1');
+    test(new Date(2010,  2), ['{Q}'], '1', 'Mar is Q1');
+    test(new Date(2010,  3), ['{Q}'], '2', 'Apr is Q2');
+    test(new Date(2010,  4), ['{Q}'], '2', 'May is Q2');
+    test(new Date(2010,  5), ['{Q}'], '2', 'Jun is Q2');
+    test(new Date(2010,  6), ['{Q}'], '3', 'Jul is Q3');
+    test(new Date(2010,  7), ['{Q}'], '3', 'Aug is Q3');
+    test(new Date(2010,  8), ['{Q}'], '3', 'Sep is Q3');
+    test(new Date(2010,  9), ['{Q}'], '4', 'Oct is Q4');
+    test(new Date(2010, 10), ['{Q}'], '4', 'Nov is Q4');
+    test(new Date(2010, 11), ['{Q}'], '4', 'Dec is Q4');
 
     d = new Date('August 5, 2010 04:03:02');
 
@@ -1536,12 +1613,12 @@ package('Date', function () {
     test(d, ['{yyyy}-{MM}-{dd} {hh}:{mm}:{ss}Z'], '2010-08-05 04:03:02Z', 'ISO8601 UTC');
     test(d, ['{Month}, {yyyy}'], 'August, 2010', 'month and year');
 
-    var isotzd = getExpectedTimezoneOffset(d, true);
+    var tz = getExpectedTimezoneOffset(d, true);
     test(d, [Sugar.Date.ISO8601_DATE], '2010-08-05', 'ISO8601_DATE | constant');
-    test(d, [Sugar.Date.ISO8601_DATETIME], '2010-08-05T04:03:02.000'+isotzd, 'ISO8601_DATETIME | constant');
+    test(d, [Sugar.Date.ISO8601_DATETIME], '2010-08-05T04:03:02.000'+tz, 'ISO8601_DATETIME | constant');
 
     test(d, ['ISO8601_DATE'], '2010-08-05', 'ISO8601_DATE | string');
-    test(d, ['ISO8601_DATETIME'], '2010-08-05T04:03:02.000'+isotzd, 'ISO8601_DATETIME | string');
+    test(d, ['ISO8601_DATETIME'], '2010-08-05T04:03:02.000'+tz, 'ISO8601_DATETIME | string');
 
     // RFC
     var d = new Date('August 5, 2010 04:03:02');
@@ -1556,8 +1633,8 @@ package('Date', function () {
     var d = run(new Date('August 5, 2010 04:03:02'), 'setUTC', [true]);
     rfc1123 = testCapitalize(getWeekdayFromDate(d,true).slice(0,3))+', '+testPadNumber(d.getUTCDate(), 2)+' '+testCapitalize(getMonthFromDate(d,true).slice(0,3))+' '+d.getUTCFullYear()+' '+testPadNumber(d.getUTCHours(), 2)+':'+testPadNumber(d.getUTCMinutes(), 2)+':'+testPadNumber(d.getUTCSeconds(), 2)+' +0000';
     rfc1036 = testCapitalize(getWeekdayFromDate(d,true))+', '+testPadNumber(d.getUTCDate(), 2)+'-'+testCapitalize(getMonthFromDate(d,true).slice(0,3))+'-'+d.getUTCFullYear().toString().slice(2)+' '+testPadNumber(d.getUTCHours(), 2)+':'+testPadNumber(d.getUTCMinutes(), 2)+':'+testPadNumber(d.getUTCSeconds(), 2)+' +0000';
-    test(d, ['RFC1123'], rfc1123, 'Date#format | string constants | RFC1123 UTC');
-    test(d, ['RFC1036'], rfc1036, 'Date#format | string constants | RFC1036 UTC');
+    test(d, ['RFC1123'], rfc1123, 'string constants | RFC1123 UTC');
+    test(d, ['RFC1036'], rfc1036, 'string constants | RFC1036 UTC');
 
     // ISO8601 - UTC
     var d = run(new Date('August 5, 2010 04:03:02'), 'setUTC', [true]);
@@ -1567,6 +1644,125 @@ package('Date', function () {
 
     test(new Date(NaN), [Sugar.Date.ISO8601_DATETIME], 'Invalid Date', 'invalid');
 
+
+    // strftime tokens
+
+    var d = new Date('August 5, 2010 14:03:02');
+
+    test(d, ['%a'], 'Thu', 'Abbreviated weekday');
+    test(d, ['%A'], 'Thursday', 'Weekday');
+    test(d, ['%b'], 'Aug', 'Abbreviated month');
+    test(d, ['%B'], 'August', 'Month');
+    test(d, ['%c'], 'Thu Aug 5 2010 2:03 PM', 'Preferred stamp');
+    test(d, ['%C'], '20', 'Century as 2 digit string');
+    test(d, ['%d'], '05', 'Padded day of the month');
+    test(d, ['%D'], '08/05/10', 'American 2 digit abbreviation');
+    test(d, ['%e'], ' 5', 'Day of the month with space padding');
+    test(d, ['%F'], '2010-08-05', 'Common datestamp format');
+    test(d, ['%g'], '10', '2 digit week year');
+    test(d, ['%G'], '2010', '4 digit week year');
+    test(d, ['%h'], 'Aug', 'Abbreviated month');
+    test(d, ['%H'], '14', 'Hours in 24 hour');
+    test(d, ['%I'], '02', 'Hours in 12 hour');
+    test(d, ['%j'], '217', 'Day of the year');
+    test(d, ['%m'], '08', '2 digit month');
+    test(d, ['%M'], '03', '2 digit minute');
+    test(d, ['%p'], 'PM', 'PM');
+    test(d, ['%P'], 'pm', 'pm');
+    test(d, ['%r'], '02:03:02 PM', '12 hour clock time with seconds');
+    test(d, ['%R'], '14:03', '24 hour clock time without seconds');
+    test(d, ['%S'], '02', 'seconds');
+    test(d, ['%T'], '14:03:02', '24 hour clock time with seconds');
+    test(d, ['%u'], '4', 'ISO8601 day of week');
+    test(d, ['%U'], '31', 'Week number with first Sunday as first week');
+    test(d, ['%V'], '31', 'ISO week number');
+    test(d, ['%w'], '4', 'Day of the week (Sunday as 0)');
+    test(d, ['%W'], '31', 'Week number with first Monday we first week');
+    test(d, ['%x'], '08/05/2010', 'Locale based representation | date only');
+    test(d, ['%X'], '2:03 PM', 'Locale based representation | time only');
+    test(d, ['%Y'], '2010', 'Full year');
+
+    equal(/[+-]\d{4}/.test(run(d, 'format', ['%z'])), true, 'Timezone offset');
+    equal(/\w{3}/.test(run(d, 'format', ['%Z'])), true, 'Timezone abbreviation');
+
+    test(new Date('January 1, 2010'), ['%c', 'en-GB'], 'Fri 1 Jan 2010 0:00', 'Preferred stamp | UK');
+
+    test(new Date('December 28, 2008'), ['%g'], '08', '2 digit week year | ISO8601');
+    test(new Date('December 28, 2008'), ['%G'], '2008', '4 digit week year | ISO8601');
+    test(new Date('December 29, 2008'), ['%g'], '09', '2 digit week year | next | ISO8601');
+    test(new Date('December 29, 2008'), ['%G'], '2009', '4 digit week year | next | ISO8601');
+    test(new Date('January 3, 2010'),   ['%g'], '09', '2 digit week year | prev | ISO8601');
+    test(new Date('January 3, 2010'),   ['%G'], '2009', '4 digit week year | prev | ISO8601');
+
+    test(new Date('August 5, 2010 4:03:02'), ['%H'], '04', 'Padded hours in 24 hour');
+    test(new Date('August 5, 2010 23:59:59'), ['%j'], '217', 'Day of the year at end of day');
+    test(new Date('August 5, 2010 00:00:00'), ['%j'], '217', 'Day of the year at beginning of day');
+    test(new Date('August 5, 2008 00:00:00'), ['%j'], '218', 'Day of the year on leap year');
+    test(new Date('January 1, 2010 00:00:00'), ['%j'], '001', 'Padded day of the year');
+    test(new Date('December 31, 2010 23:59:59'), ['%j'], '365', 'Last second of the year');
+    test(new Date('August 5, 2010 4:00'), ['%p'], 'AM', 'AM');
+    test(new Date('August 5, 2010 4:00'), ['%P'], 'am', 'am');
+
+    test(new Date('January 3, 2010'), ['%u'], '7', 'ISO8601 day of week Sunday');
+    test(new Date('January 4, 2010'), ['%u'], '1', 'ISO8601 day of week Monday');
+
+    test(new Date('January 3, 2010'), ['%w'], '0', 'Day of week Sunday');
+    test(new Date('January 4, 2010'), ['%w'], '1', 'Day of week Monday');
+
+    test(new Date('January 1, 2010'), ['%U'], '00', 'Week of the year (Sunday first) 1st');
+    test(new Date('January 2, 2010'), ['%U'], '00', 'Week of the year (Sunday first) 2nd');
+    test(new Date('January 3, 2010'), ['%U'], '01', 'Week of the year (Sunday first) 3rd');
+    test(new Date('January 4, 2010'), ['%U'], '01', 'Week of the year (Sunday first) 4th');
+
+    test(new Date('January 1, 2010'), ['%V'], '53', 'ISO week of the year 1st');
+    test(new Date('January 2, 2010'), ['%V'], '53', 'ISO week of the year 2nd');
+    test(new Date('January 3, 2010'), ['%V'], '53', 'ISO week of the year 3rd');
+    test(new Date('January 4, 2010'), ['%V'], '01', 'ISO week of the year 4th');
+
+    test(new Date('January 1, 2010'), ['%W'], '00', 'Week of the year (Monday first) 1st');
+    test(new Date('January 2, 2010'), ['%W'], '00', 'Week of the year (Monday first) 2nd');
+    test(new Date('January 3, 2010'), ['%W'], '00', 'Week of the year (Monday first) 3rd');
+    test(new Date('January 4, 2010'), ['%W'], '01', 'Week of the year (Monday first) 4th');
+
+    test(new Date('May 1, 2010'), ['%b'], 'May', 'May');
+
+    test(new Date(-60000000000000), ['%y'], '68', '%y in year 0068 should be 2 digits');
+    test(new Date(-60000000000000), ['%Y'], '0068', '%Y in year 0068 should be 4 digits');
+
+    test(d, ['%Y%A'], '2010Thursday', 'tokens together');
+    test(d, ['%Y\n%A'], '2010\nThursday', 'can input newlines');
+    test(d, ['%Y\t%A'], '2010\tThursday', 'can input tabs');
+    test(d, ['%Y%%%A'], '2010%Thursday', 'can escape percent signs');
+
+
+    // Shortcuts
+
+    var then = new Date(2010, 0, 5, 15, 52);
+
+    assertFormatShortcut(then, 'short', '01/05/2010');
+    assertFormatShortcut(then, 'medium', 'January 5, 2010');
+    assertFormatShortcut(then, 'long', 'January 5, 2010 3:52 PM');
+    assertFormatShortcut(then, 'full', 'Tuesday, January 5, 2010 3:52 PM');
+    test(then, ['{time}'], '3:52 PM', 'preferred time');
+    test(then, ['{stamp}'], 'Tue Jan 5 2010 3:52 PM', 'preferred stamp');
+
+    assertFormatShortcut(then, 'short', '05/01/2010', 'en-GB');
+    assertFormatShortcut(then, 'medium', '5 January 2010', 'en-GB');
+    assertFormatShortcut(then, 'long', '5 January 2010 15:52', 'en-GB');
+    assertFormatShortcut(then, 'full', 'Tuesday, 5 January, 2010 15:52', 'en-GB');
+    test(then, ['{time}', 'en-GB'], '15:52', 'preferred time | UK');
+    test(then, ['{stamp}', 'en-GB'], 'Tue 5 Jan 2010 15:52', 'preferred stamp | UK');
+
+    assertFormatShortcut(then, 'short', '05/01/2010', 'en-AU');
+    assertFormatShortcut(then, 'medium', '5 January 2010', 'en-AU');
+    assertFormatShortcut(then, 'long', '5 January 2010 15:52', 'en-AU');
+    assertFormatShortcut(then, 'full', 'Tuesday, 5 January, 2010 15:52', 'en-AU');
+
+    assertFormatShortcut(then, 'short', '2010-01-05', 'en-CA');
+    assertFormatShortcut(then, 'medium', '5 January, 2010', 'en-CA');
+    assertFormatShortcut(then, 'long', '5 January, 2010 15:52', 'en-CA');
+    assertFormatShortcut(then, 'full', 'Tuesday, 5 January, 2010 15:52', 'en-CA');
+
     // Issue #262
     equal(/\d+-/.test(run(new Date(), 'format', ['{timezone}'])), false, 'Timezone format should not include hyphens')
 
@@ -1575,17 +1771,6 @@ package('Date', function () {
     test(new Date(1901, 0, 2), ['{yyyy}'], '1901', 'Zero padded year should respect yyyy format');
 
   });
-
-  group('Locale Specific Formats', function() {
-    var d = new Date('August 5, 2010 04:03:02');
-    equal(run(d, 'format', ['short']), 'August 5, 2010', 'Date#format | shortcuts | short');
-    equal(run(d, 'short'), 'August 5, 2010', 'Date#format | shortcuts | short method');
-    equal(run(d, 'format', ['long']), 'August 5, 2010 4:03am', 'Date#format | shortcuts | long');
-    equal(run(d, 'long'), 'August 5, 2010 4:03am', 'Date#format | shortcuts | long method');
-    equal(run(d, 'format', ['full']), 'Thursday August 5, 2010 4:03:02am', 'Date#format | shortcuts | full');
-    equal(run(d, 'full'), 'Thursday August 5, 2010 4:03:02am', 'Date#format | shortcuts | full method');
-  });
-
 
   method('getUTCOffset', function() {
     var d = new Date('August 5, 2010 04:03:02');
@@ -1651,7 +1836,7 @@ package('Date', function () {
       equal(unit, 0, 'still passes millisecond is zero');
     }]);
 
-    equal(run(testCreateDate('2002-02-17'), 'format', [function() {}]), 'February 17, 2002 12:00am', 'function that returns undefined defaults to standard format');
+    equal(run(testCreateDate('2002-02-17'), 'format', [function() {}]), 'February 17, 2002 12:00 AM', 'function that returns undefined defaults to standard format');
 
     equal(run(testCreateDate(), 'relative'), '1 second ago', '6 milliseconds ago');
 
@@ -1876,33 +2061,6 @@ package('Date', function () {
     test(testCreateDate('1800'), false, '1800');
     test(testCreateDate('1900'), false, '1900');
     test(testCreateDate('2000'), true, '2000');
-  });
-
-  method('getISOWeek', function() {
-    var d;
-
-    d = new Date(2010,7,5,13,45,2,542);
-
-    test(d, 31, 'basic August 5th, 2010');
-    dateEqual(d, new Date(2010,7,5,13,45,2,542), 'should not modify the date');
-
-    test(new Date(2010, 0, 1), 53, 'January 1st, 2010');
-    test(new Date(2010, 0, 6), 1, 'January 6th, 2010');
-    test(new Date(2010, 0, 7), 1, 'January 7th, 2010');
-    test(new Date(2010, 0, 7, 23, 59, 59, 999), 1, 'January 7th, 2010 h23:59:59.999');
-    test(new Date(2010, 0, 8), 1, 'January 8th, 2010');
-    test(new Date(2010, 3, 15), 15, 'April 15th, 2010');
-
-    d = run(new Date(2010,7,5,13,45,2,542), 'setUTC', [true]);
-
-    test(d, d.getTimezoneOffset() > 615 ? 32 : 31, 'utc | basic');
-    test(new Date(2010, 0, 1), 53, 'utc | January 1st UTC is actually 2009');
-    test(new Date(2010, 0, 6), 1, 'utc | January 6th');
-    test(new Date(2010, 0, 7), 1, 'utc | January 7th');
-    test(new Date(2010, 0, 7, 23, 59, 59, 999), 1, 'utc | January 7th 23:59:59.999');
-    test(new Date(2010, 0, 8), 1, 'utc | January 8th');
-    test(new Date(2010, 3, 15), 15, 'utc | April 15th');
-
   });
 
   group('Since/Until', function() {
@@ -2624,7 +2782,7 @@ package('Date', function () {
     raisesError(function(){ testSetLocale(); }, 'no arguments raises error');
     equal(testGetLocale().code, 'fo', 'setting locale with no arguments had no effect');
     equal(run(new Date(2011, 5, 6), 'format', ['{Month}']), 'La', 'will not change the locale if no argument passed');
-    equal(run(new Date(2011, 5, 6), 'format', ['', 'en']), 'June 6, 2011 12:00am', 'local locale should override global');
+    equal(run(new Date(2011, 5, 6), 'format', ['', 'en']), 'June 6, 2011 12:00 AM', 'local locale should override global');
     equal(run(testCreateDate('5 months ago', 'en'), 'relative', ['en']), '5 months ago', 'local locale should override global');
 
     raisesError(function(){ testSetLocale(''); }, '"" raises an invalid locale error');
@@ -2637,6 +2795,10 @@ package('Date', function () {
   });
 
   method('getISOWeek', function() {
+
+    var d = new Date(2010,7,5,13,45,2,542);
+    test(d, 31, 'basic August 5th, 2010');
+    dateEqual(d, new Date(2010,7,5,13,45,2,542), 'should not modify the date');
 
     test(new Date(2011, 0, 1), 52, 'January 1, 2011');
     test(new Date(2011, 0, 2), 52, 'January 2, 2011');
@@ -2659,6 +2821,22 @@ package('Date', function () {
     test(new Date(2014,  0,  2),  1, 'January 02, 2014');
     test(new Date(2014,  0,  5),  1, 'January 05, 2014');
     test(new Date(2014,  0,  6),  2, 'January 06, 2014');
+
+    test(new Date(2010, 0, 1), 53, 'January 1st, 2010');
+    test(new Date(2010, 0, 6), 1, 'January 6th, 2010');
+    test(new Date(2010, 0, 7), 1, 'January 7th, 2010');
+    test(new Date(2010, 0, 7, 23, 59, 59, 999), 1, 'January 7th, 2010 h23:59:59.999');
+    test(new Date(2010, 0, 8), 1, 'January 8th, 2010');
+    test(new Date(2010, 3, 15), 15, 'April 15th, 2010');
+
+    var d = run(new Date(2010,7,5,13,45,2,542), 'setUTC', [true]);
+    test(d, d.getTimezoneOffset() > 615 ? 32 : 31, 'utc | basic');
+    test(new Date(2010, 0, 1), 53, 'utc | January 1st UTC is actually 2009');
+    test(new Date(2010, 0, 6), 1, 'utc | January 6th');
+    test(new Date(2010, 0, 7), 1, 'utc | January 7th');
+    test(new Date(2010, 0, 7, 23, 59, 59, 999), 1, 'utc | January 7th 23:59:59.999');
+    test(new Date(2010, 0, 8), 1, 'utc | January 8th');
+    test(new Date(2010, 3, 15), 15, 'utc | April 15th');
 
   });