+++ test/tests/locales/ko.js
@@ -1,10 +1,11 @@
 package('Dates Korean', function () {
   "use strict";
 
-  var now;
+  var now, then;
 
   setup(function() {
     now = new Date();
+    then = new Date(2010, 0, 5, 15, 52);
     testSetLocale('ko');
   });
 
@@ -75,17 +76,45 @@ package('Dates Korean', function () {
   });
 
   method('format', function() {
-    var then = new Date(2011, 7, 25, 15, 45, 50);
-    test(then, '2011년8월25일 15시45분', 'standard format');
-    test(then, ['{yyyy}년{MM}월{dd}일'], '2011년08월25일', 'format');
 
-    // Format shortcuts
-    equal(run(then, 'format', ['full']), '2011년8월25일 목요일 15시45분50초', 'full format');
-    equal(run(then, 'full'), '2011년8월25일 목요일 15시45분50초', 'full shortcut');
-    equal(run(then, 'format', ['long']), '2011년8월25일 15시45분', 'long format');
-    equal(run(then, 'long'), '2011년8월25일 15시45분', 'long shortcut');
-    equal(run(then, 'format', ['short']), '2011년8월25일', 'short format');
-    equal(run(then, 'short'), '2011년8월25일', 'short shortcut');
+    test(then, '2010년 1월 5일 오후 3시 52분', 'default format');
+
+    assertFormatShortcut(then, 'short', '2010.01.05');
+    assertFormatShortcut(then, 'medium', '2010년 1월 5일');
+    assertFormatShortcut(then, 'long', '2010년 1월 5일 오후 3시 52분');
+    assertFormatShortcut(then, 'full', '2010년 1월 5일 화요일 오후 3시 52분');
+    test(then, ['{time}'], '오후 3시 52분', 'preferred time');
+    test(then, ['{stamp}'], '2010년 1월 5일 15:52 화', 'preferred stamp');
+    test(then, ['%c'], '2010년 1월 5일 15:52 화', '%c stamp');
+
+    test(new Date('January 3, 2010'), ['{w}'], '53', 'locale week number | Jan 3 2010');
+    test(new Date('January 3, 2010'), ['{ww}'], '53', 'locale week number padded | Jan 3 2010');
+    test(new Date('January 3, 2010'), ['{wo}'], '53rd', 'locale week number ordinal | Jan 3 2010');
+    test(new Date('January 4, 2010'), ['{w}'], '1', 'locale week number | Jan 4 2010');
+    test(new Date('January 4, 2010'), ['{ww}'], '01', 'locale week number padded | Jan 4 2010');
+    test(new Date('January 4, 2010'), ['{wo}'], '1st', 'locale week number ordinal | Jan 4 2010');
+
+    test(new Date(2015, 10, 8),  ['{Dow}'], '일', 'Sun');
+    test(new Date(2015, 10, 9),  ['{Dow}'], '월', 'Mon');
+    test(new Date(2015, 10, 10), ['{Dow}'], '화', 'Tue');
+    test(new Date(2015, 10, 11), ['{Dow}'], '수', 'Wed');
+    test(new Date(2015, 10, 12), ['{Dow}'], '목', 'Thu');
+    test(new Date(2015, 10, 13), ['{Dow}'], '금', 'Fri');
+    test(new Date(2015, 10, 14), ['{Dow}'], '토', 'Sat');
+
+    test(new Date(2015, 0, 1),  ['{Mon}'], '1월', 'Jan');
+    test(new Date(2015, 1, 1),  ['{Mon}'], '2월', 'Feb');
+    test(new Date(2015, 2, 1),  ['{Mon}'], '3월', 'Mar');
+    test(new Date(2015, 3, 1),  ['{Mon}'], '4월', 'Apr');
+    test(new Date(2015, 4, 1),  ['{Mon}'], '5월', 'May');
+    test(new Date(2015, 5, 1),  ['{Mon}'], '6월', 'Jun');
+    test(new Date(2015, 6, 1),  ['{Mon}'], '7월', 'Jul');
+    test(new Date(2015, 7, 1),  ['{Mon}'], '8월', 'Aug');
+    test(new Date(2015, 8, 1),  ['{Mon}'], '9월', 'Sep');
+    test(new Date(2015, 9, 1),  ['{Mon}'], '10월', 'Oct');
+    test(new Date(2015, 10, 1), ['{Mon}'], '11월', 'Nov');
+    test(new Date(2015, 11, 1), ['{Mon}'], '12월', 'Dec');
+
   });
 
 
@@ -119,6 +148,11 @@ package('Dates Korean', function () {
     assertRelative('5 days from now', '5일 후');
     assertRelative('5 weeks from now', '1개월 후');
     assertRelative('5 years from now', '5년 후');
+  });
+
+  method('beginning/end', function() {
+    dateEqual(dateRun(new Date(2010, 0), 'beginningOfWeek'), new Date(2009, 11, 28), 'beginningOfWeek');
+    dateEqual(dateRun(new Date(2010, 0), 'endOfWeek'), new Date(2010, 0, 3, 23, 59, 59, 999), 'endOfWeek');
   });
 
 });