+++ test/tests/locales/sv.js
@@ -1,10 +1,11 @@
 package('Dates Swedish', function () {
   "use strict";
 
-  var now;
+  var now, then;
 
   setup(function() {
     now = new Date();
+    then = new Date(2010, 0, 5, 15, 52);
     testSetLocale('sv');
   });
 
@@ -104,8 +105,45 @@ package('Dates Swedish', function () {
   });
 
   method('format', function() {
-    test(testCreateDate('2001-06-14 3:45pm'), 'den 14 juni 2001 15:45', 'default format');
-    test(testCreateDate('2011-08-25'), ['{dd} {month} {yyyy}'], '25 augusti 2011', 'dd month yyyy format');
+
+    test(then, '5 januari 2010 15:52', 'default format');
+
+    assertFormatShortcut(then, 'short', '2010-01-05');
+    assertFormatShortcut(then, 'medium', '5 januari 2010');
+    assertFormatShortcut(then, 'long', '5 januari 2010 15:52');
+    assertFormatShortcut(then, 'full', 'tisdag 5 januari 2010 15:52');
+    test(then, ['{time}'], '15:52', 'preferred time');
+    test(then, ['{stamp}'], 'tis 5 jan 2010 15:52', 'preferred stamp');
+    test(then, ['%c'], 'tis 5 jan 2010 15:52', '%c stamp');
+
+    test(new Date('January 3, 2010'), ['{w}'], '53', 'locale week number | Jan 3 2010');
+    test(new Date('January 3, 2010'), ['{ww}'], '53', 'locale week number padded | Jan 3 2010');
+    test(new Date('January 3, 2010'), ['{wo}'], '53rd', 'locale week number ordinal | Jan 3 2010');
+    test(new Date('January 4, 2010'), ['{w}'], '1', 'locale week number | Jan 4 2010');
+    test(new Date('January 4, 2010'), ['{ww}'], '01', 'locale week number padded | Jan 4 2010');
+    test(new Date('January 4, 2010'), ['{wo}'], '1st', 'locale week number ordinal | Jan 4 2010');
+
+    test(new Date(2015, 10, 8),  ['{Dow}'], 'sön', 'Sun');
+    test(new Date(2015, 10, 9),  ['{Dow}'], 'mån', 'Mon');
+    test(new Date(2015, 10, 10), ['{Dow}'], 'tis', 'Tue');
+    test(new Date(2015, 10, 11), ['{Dow}'], 'ons', 'Wed');
+    test(new Date(2015, 10, 12), ['{Dow}'], 'tor', 'Thu');
+    test(new Date(2015, 10, 13), ['{Dow}'], 'fre', 'Fri');
+    test(new Date(2015, 10, 14), ['{Dow}'], 'lör', 'Sat');
+
+    test(new Date(2015, 0, 1),  ['{Mon}'], 'jan', 'Jan');
+    test(new Date(2015, 1, 1),  ['{Mon}'], 'feb', 'Feb');
+    test(new Date(2015, 2, 1),  ['{Mon}'], 'mar', 'Mar');
+    test(new Date(2015, 3, 1),  ['{Mon}'], 'apr', 'Apr');
+    test(new Date(2015, 4, 1),  ['{Mon}'], 'maj', 'May');
+    test(new Date(2015, 5, 1),  ['{Mon}'], 'jun', 'Jun');
+    test(new Date(2015, 6, 1),  ['{Mon}'], 'jul', 'Jul');
+    test(new Date(2015, 7, 1),  ['{Mon}'], 'aug', 'Aug');
+    test(new Date(2015, 8, 1),  ['{Mon}'], 'sep', 'Sep');
+    test(new Date(2015, 9, 1),  ['{Mon}'], 'okt', 'Oct');
+    test(new Date(2015, 10, 1), ['{Mon}'], 'nov', 'Nov');
+    test(new Date(2015, 11, 1), ['{Mon}'], 'dec', 'Dec');
+
   });
 
 
@@ -141,6 +179,11 @@ package('Dates Swedish', function () {
     assertRelative('3 weeks from now',   'om 3 veckor');
     assertRelative('5 weeks from now',   'om 1 månad');
     assertRelative('5 year from now',    'om 5 år');
+  });
+
+  method('beginning/end', function() {
+    dateEqual(dateRun(new Date(2010, 0), 'beginningOfWeek'), new Date(2009, 11, 28), 'beginningOfWeek');
+    dateEqual(dateRun(new Date(2010, 0), 'endOfWeek'), new Date(2010, 0, 3, 23, 59, 59, 999), 'endOfWeek');
   });
 
 });